refactor(render): extract CmdCategory into its own module
